refactor: 채팅방 TopBar 컴포넌트로 분리
